Predicates implementation (#89)
* Added builtin predicates
* Added and predicate.
* Added missing methods in ObjectDataOutput and ObjectDataInput for read/writeObject and added BetweenPredicate.
* EqualPredicate
* GreaterLessPredicate
* LikePredicate
* InPredicate
* InstanceOfPredicate
* Not and NotEqual
* OrPredicate
* RegExPredicate
* Added Predicate interface, ClientMapTest not compiling due to PagingPredicate.
* PagingPredicate
* Predicate with EntryListener is added.
* Added values with predicate test
* testEntrySetWithPredicateTest and testKeySetWithPredicateTest added
* RawPointerMap API is corrected.
* Changed the DataArray and EntryArray interfaces. Made the PagingPredicate work for primitives.
* Fix for setAnchor
* Update To map test and raw pointer examples
* Added example for builtin predicate queries.
* Added entry listener with different predicates tests.

+++ hazelcast/test/src/util/ClientUtilTest.cpp
@@ -69,7 +69,7 @@ namespace hazelcast {
                     if (wokenUpByInterruption) {
                         end = time(NULL);
                     }
-                    ASSERT_NEAR(end-beg, wakeUpTime, 1);
+                    ASSERT_NEAR((double)(end-beg), (double)wakeUpTime, 1);
                 }
 
             }
@@ -84,7 +84,7 @@ namespace hazelcast {
                 } catch (exception::TimeoutException&) {
                     end = time(NULL);
                 }
-                ASSERT_NEAR(end-beg, waitSeconds, 1);
+                ASSERT_NEAR((double)(end-beg), (double)waitSeconds, 1);
             }
 
             TEST_F (ClientUtilTest, testFutureSetValue) {
@@ -155,7 +155,7 @@ namespace hazelcast {
                 thread.cancel();
                 thread.join();
                 time_t end = time(NULL);
-                ASSERT_NEAR((int)(end - beg), wakeUpTime , 1);
+                ASSERT_NEAR((double)(end - beg), (double)wakeUpTime , 1);
             }
         }
     }
